Decouple test fixtures from techniques config yaml
Move PCA-specific tests from test_simulation.py and test_prediction.py
into test_pca.py with local fixtures that hardcode
DimensionalityReductionPCA, bypassing the conftest cross-product. Bundle
(technique, comp) pairs in conftest's dr_technique fixture so each DR
class gets an appropriate comp value — fixes KernelPCA
InvalidParameterError from receiving a float. Remove unused
dummy_sim_array fixture from test_simulation.py. Use tmp_path fixture
for test output directories. Update test_pca based on new prepare
function signature.
