+++ tests/test_prediction.py
@@ -240,46 +240,3 @@ def test_train_test_series_standard_case(setup_prediction_class):
     total_len = steps
     expected_x_pred = np.linspace(1 + pas, 1 + steps * pas, steps).reshape(-1, 1)
     assert np.allclose(x_pred, expected_x_pred)
-
-
-CASES = [
-    ("ssh", "DINO_1m_To_1y_grid_T.nc"),
-    ("soce", "DINO_1y_grid_T.nc"),
-    ("toce", "DINO_1y_grid_T.nc"),
-]
-
-# Duplicate each case so BOTH fixtures receive the same tuple
-PARAM_ROWS = [pytest.param(c, c, id=f"{c[0]}-{c[1]}") for c in CASES]
-
-
-@pytest.mark.parametrize(
-    ("setup_prediction_class", "setup_simulation_class"),
-    PARAM_ROWS,
-    indirect=["setup_prediction_class", "setup_simulation_class"],
-)
-def test_predictions_reconstruct(setup_prediction_class, setup_simulation_class):
-    """
-    Check that the reconstruct function rebuilds the time series correctly.
-
-    This will check the result is the correct shape.
-    """
-    # setup prediction class
-    pred, infos = setup_prediction_class
-    sim = setup_simulation_class
-
-    steps = 20
-
-    # Forecast specified number of steps
-    y_hat, _y_hat_std, _metrics = pred.parallel_forecast(len(pred), steps)
-
-    # Reconstruct with n predicted components
-    n = len(pred.info["pca"].components_)
-    reconstructed_preds = sim.reconstruct(
-        y_hat, n, infos, begin=0
-    )  # TODO: Use simulation prediction class, setup simulation class fixture
-
-    # Expected shape: forecast steps x original spatial dimensions
-    expected_shape = (steps, *tuple(pred.info["shape"]))
-    assert reconstructed_preds.shape == expected_shape, (
-        f"Reconstructed shape {reconstructed_preds.shape} != expected {expected_shape}"
-    )
